fix: Loading issue with ModesLists and PurposesList
We have the same checks between Modes & Purposes so let's use an Outlet
in the `/analysis` route to handle it only once. Like that we avoid to
refresh the DateSelection and else.

Also take the same refacto than the TripsView to have a better management
of the empty / loading / loaded state
